refactor(tests): consolidate bitcoin.signtx helper module
